api/requirements.txt
asgiref==3.3.1
Django==3.0.0
django-cors-headers==3.5.0
djangorestframework==3.12.2
Markdown==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pytesseract==0.3.6
pytz==2020.4
sqlparse==0.4.1